Move PluginAssemblyResolver to SpecFlow
