+++ TechTalk.SpecFlow.Generator/TechTalk.SpecFlow.Generator.csproj
@@ -43,7 +43,6 @@
   
   <ItemGroup Condition="'$(TargetFramework)' == '$(SpecFlow_Core_Generator_TFM)' Or '$(TargetFramework)' == '$(SpecFlow_Net5_TFM)'">
     <PackageReference Include="System.Configuration.ConfigurationManager" Version="4.4.1" />
-    <PackageReference Include="Microsoft.Extensions.DependencyModel" Version="1.0.3" />
     <PackageReference Include="System.Net.Http" Version="4.3.4" />
     <PackageReference Include="System.Collections" Version="4.3.0" />
     <PackageReference Include="System.Diagnostics.Debug" Version="4.3.0" />
