+++ TechTalk.SpecFlow/SpecFlow.nuspec
@@ -27,6 +27,8 @@
         <dependency id="Utf8Json" version="1.3.7" />
         <dependency id="System.ValueTuple" version="4.4.0" />
         <dependency id="System.Configuration.ConfigurationManager" version="4.5.0" />
+        <dependency id="Microsoft.Extensions.DependencyModel" version="1.0.3" />
+        <dependency id="System.Runtime.Loader" version="4.3.0" />
         <dependency id="BoDi" version="1.5.0" />
         <dependency id="Gherkin" version="6.0.0"/>
         <dependency id="Cucumber.Messages" version="[6.0.1, 7.0)" />
